backend/requirements.txt
# Dependencias principales del backend
fastapi==0.104.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
pydantic==2.5.0
python-multipart==0.0.6

# Dependencias para comunicación con Ollama
requests==2.31.0

# Sistema RAG - Dependencias de IA y ML
sentence-transformers==2.2.2
pinecone==3.0.3
pypdf==3.17.4
numpy==1.24.3
transformers==4.35.2
torch==2.1.1

# Dependencias adicionales para procesamiento
scikit-learn==1.3.2
tqdm==4.66.1